File Encoding
Update all functions using read.csv or write.csv to force fileEncoding = "ISO-8859-1".
This means that all csv files in BPWTP_labdat will be and will have to be ISO-8859-1.
This solves encoding problems where UTF-8 and ISO-8859-1 files were not being handled the same way on all systems. This had created problems with units and the mu µ symbol that broke the monthly report in several places.